+++ InscryptionAPI/Triggers/Interfaces.cs
@@ -802,6 +802,24 @@ public interface IOnCardDealtDamageDirectly
     public IEnumerator OnCardDealtDamageDirectly(PlayableCard attacker, CardSlot opposingSlot, int damage);
 }
 
+/// <summary>
+/// Trigger that is called after a shielded card was attacked and lost a shield.
+/// </summary>
+public interface IShieldPreventedDamage
+{
+    public bool RespondsToShieldPreventedDamage(PlayableCard target, int damage, PlayableCard attacker);
+    public IEnumerator OnShieldPreventedDamage(PlayableCard target, int damage, PlayableCard attacker);
+    public int ShieldPreventedDamagePriority(PlayableCard target, int damage, PlayableCard attacker);
+}
+/// <summary>
+/// Variant of IShieldPreventDamage that triggers for cards in the hand.
+/// </summary>
+public interface IShieldPreventedDamageInHand
+{
+    public bool RespondsToShieldPreventedDamageInHand(PlayableCard target, int damage, PlayableCard attacker);
+    public IEnumerator OnShieldPreventedDamageInHand(PlayableCard target, int damage, PlayableCard attacker);
+    public int ShieldPreventedDamageInHandPriority(PlayableCard target, int damage, PlayableCard attacker);
+}
 /*public interface IOnPreTakeDamageFromHammer
 {
     public bool RespondsToPreTakeDamageFromHammer(HammerItem hammer, CardSlot targetSlot, GameObject firstPersonItem);